pyproject.toml
[project]
name = "agentos-runtime"
version = "1.0.0"
description = "Distributed LLM Agent Orchestration Runtime — FastAPI + LangGraph + Redis Streams + PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn[standard]>=0.44.0",
    "langgraph>=1.1.6",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "redis>=7.4.0",
    "asyncpg>=0.31.0",
    "sqlalchemy>=2.0.49",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.13.1",
    "websockets>=16.0",
    "python-dotenv>=1.2.2",
    "aiosqlite>=0.22.1",
]

[project.scripts]
agentos = "src.api.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
